backend/requirements.txt
# Django Core
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-guardian==2.4.0
PyJWT==2.8.0

# Cache
redis==5.0.1
django-redis==5.4.0

# Async & Celery (for background tasks)
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Documentation
drf-spectacular==0.27.0

# Email
django-anymail==10.2

# File Handling
Pillow==10.1.0
python-magic==0.4.27

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Monitoring & Logging
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
